requirements.txt
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
dominate==2.6.0
Flask==2.1.1
Flask-Bootstrap==3.3.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
ldap3==2.9.1
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.1.1
python-dateutil==2.8.2
python-ldap @ file:///D:/Descargas/python_ldap-3.4.0-cp310-cp310-win_amd64.whl
pytz==2022.1
six==1.16.0
visitor==0.1.3
Werkzeug==2.1.1
